[Feat]: 컴포넌트 분리, 네비게이션 연결기능 구현
