+++ feature/myaccount/src/main/java/com/android/myfooddiarybookaos/myaccount/navi/MyScreenRoot.kt
@@ -1,0 +1,6 @@
+package com.android.myfooddiarybookaos.myaccount.navi
+
+object MyScreenRoot {
+    const val MY = "myMainScreen"
+    const val NOTICE = "noticeScreen"
+}
